+++ src/interfaces/ISignedItem.ts
@@ -1,9 +1,0 @@
-/**
- * Represents an item of data stored with a signature.
- * @interface
- */
-export interface ISignedItem<T> {
-    data: T;
-    timestamp: number;
-    signature: string;
-}