+++ test/storage/storageClient.spec.ts
@@ -11,7 +11,7 @@ describe("StorageClient", () => {
     let sandbox: sinon.SinonSandbox;
 
     beforeEach(() => {
-        sandbox = sinon.sandbox.create();
+        sandbox = sinon.createSandbox();
         transactionClientStub = <ITransactionClient>{};
     });